Handle case where anchor not on expected page
amovsisyan's patch looks for an anchor to save the position relative to.
Which works fine as long as there are anchors on every page.
Unfortunately with tables this often is not the case.

If GetDocumentPositionByXY() returns an anchor on a previous, invisible
page, try moving down the current page 10 pixels at a time until we find
an anchor on this page. This is similar to what the code above does.

If that does not work either, we need to reference an anchor on a
previous page. Use ConvertCoordsToAnotherPage() to do this correctly.